requirements.txt
asttokens==2.0.5
certifi==2021.5.30
colorama==0.4.4
cycler==0.10.0
executing==0.7.0
icecream==2.1.1
install==1.3.4
kiwisolver==1.3.1
matplotlib==3.4.2
MouseInfo==0.1.3
numpy==1.21.1
opencv-contrib-python==4.5.3.56
pandas==1.3.1
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.9.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.8
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2021.1
six==1.16.0
wincertstore==0.2
